backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
neo4j==5.15.0
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.1.0
scispacy==0.5.3
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.3/en_ner_bionlp13cg_md-0.5.3.tar.gz
transformers==4.36.0
torch==2.1.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
colpali-engine==0.2.0
numpy==1.24.3
pandas==2.1.4
python-multipart==0.0.6
boto3==1.34.0
minio==7.2.0
langchain==0.0.350
openai==1.3.6
httpx==0.25.2
python-dotenv==1.0.0
Jinja2==3.1.2
aiofiles==23.2.1